requirements.txt
streamlit==1.47.0
streamlit-feedback==0.1.4
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-groq==0.3.6
biopython==1.85
dotenv==0.9.9
ipykernel==6.29.5
ipython==9.4.0
numpy==2.3.1
pandas==2.3.1
xlsxwriter==3.2.5
networkx==3.5
pyvis==0.3.2
dotenv==0.9.9
md2docx-python==1.0.0
anthropic>=0.3.0
trubrics>=1.4.3
duckduckgo_search==8.1.1
openai==1.97.0